E-filing HVUT Form 2290 and Form 8849 Schedule 6
Case of e-filing HVUT Form 2290 and Form 8849 Schedule 6 together
Have you ever come across the scenario where you have to file form 2290 to report and pay HVUT to IRS and at the same time claim credit for some vehicle(s)? You must be wondering whether to file a separate Form 2290 and Form 8849 to pay HVUT and claim credits respectively. This means that you first pay the taxes for the taxable vehicles and then file form 8849 separately to claim credit. There should be a simple way to handle this case so that the credit amount is adjusted with the tax amount that you owe so that you only pay or receive the differential amount.
Taxseer2290 has a very easy and flexible solution for you that work under multiple scenarios. The following scenarios will help to understand our solution in detail: 
For More Info:  http://blog.taxseer2290.com/case-of-e-filing-hvut-form-2290-and-form-8849-schedule-6-together/
Address: 12502 Grand Elm St, Clarksburg, MD, US 20871
Use the phone # - 240.780.6153